Rename project source module to repository
