+++ packages/ploys-cli/src/project/info.rs
@@ -1,7 +1,7 @@
 use anyhow::Error;
 use clap::Args;
 use console::style;
-use ploys::project::source::revision::Revision;
+use ploys::project::repository::revision::Revision;
 use ploys::project::Project;
 
 use crate::util::repo_or_url::RepoOrUrl;
