+++ packages/ploys/src/project/mod.rs
@@ -36,7 +36,7 @@
 
 mod error;
 mod file;
-pub mod source;
+pub mod repository;
 
 use std::path::{Path, PathBuf};
 
@@ -48,7 +48,7 @@ use crate::package::{Bump, Package};
 
 pub use self::error::Error;
 pub use self::file::{File, Fileset};
-use self::source::Source;
+use self::repository::Source;
 
 /// A project from one of several supported sources.
 pub struct Project {
@@ -64,7 +64,7 @@ impl Project {
     where
         P: AsRef<Path>,
     {
-        use self::source::git::Git;
+        use self::repository::git::Git;
 
         let source = Source::Git(Git::new(path)?);
         let name = source.get_name()?;
@@ -83,9 +83,9 @@ impl Project {
     pub fn git_with_revision<P, R>(path: P, revision: R) -> Result<Self, Error>
     where
         P: AsRef<Path>,
-        R: Into<self::source::revision::Revision>,
+        R: Into<self::repository::revision::Revision>,
     {
-        use self::source::git::Git;
+        use self::repository::git::Git;
 
         let source = Source::Git(Git::new(path)?.with_revision(revision));
         let name = source.get_name()?;
@@ -108,7 +108,7 @@ impl Project {
     where
         R: AsRef<str>,
     {
-        use self::source::github::GitHub;
+        use self::repository::github::GitHub;
 
         let source = Source::GitHub(GitHub::new(repository)?.validated()?);
         let name = source.get_name()?;
@@ -127,9 +127,9 @@ impl Project {
     pub fn github_with_revision<R, V>(repository: R, revision: V) -> Result<Self, Error>
     where
         R: AsRef<str>,
-        V: Into<self::source::revision::Revision>,
+        V: Into<self::repository::revision::Revision>,
     {
-        use self::source::github::GitHub;
+        use self::repository::github::GitHub;
 
         let source = Source::GitHub(
             GitHub::new(repository)?
@@ -154,7 +154,7 @@ impl Project {
         R: AsRef<str>,
         T: Into<String>,
     {
-        use self::source::github::GitHub;
+        use self::repository::github::GitHub;
 
         let source = Source::GitHub(
             GitHub::new(repository)?
@@ -182,10 +182,10 @@ impl Project {
     ) -> Result<Self, Error>
     where
         R: AsRef<str>,
-        V: Into<self::source::revision::Revision>,
+        V: Into<self::repository::revision::Revision>,
         T: Into<String>,
     {
-        use self::source::github::GitHub;
+        use self::repository::github::GitHub;
 
         let source = Source::GitHub(
             GitHub::new(repository)?
@@ -361,7 +361,7 @@ impl Project {
         #[cfg(feature = "github")]
         #[allow(irrefutable_let_patterns)]
         if let Source::GitHub(github) = &mut self.source {
-            use self::source::revision::{Reference, Revision};
+            use self::repository::revision::{Reference, Revision};
 
             let sha = github.commit(message, files.into_iter())?;
